+++ api/baeuja/.gitignore
@@ -44,3 +44,5 @@ out/
 .env
 src/main/resources/application.yml
 src/main/resources/application-*.yml
+src/test/resources/application.yml
+src/test/resources/application-*.yml
