+++ api/baeuja/src/main/java/xyz/baeuja/api/global/exception/BaseException.java
@@ -1,2 +1,20 @@
-package xyz.baeuja.api.global.exception;public class BaseException {
+package xyz.baeuja.api.global.exception;
+
+public class BaseException extends RuntimeException {
+
+    public BaseException() {
+        super();
+    }
+
+    public BaseException(String message) {
+        super(message);
+    }
+
+    public BaseException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public BaseException(Throwable cause) {
+        super(cause);
+    }
 }
